Create wgpu rendering instances and connect Android SurfaceView.
Created a wgpu rendering instance using Vulkan as the backend.
What is important to note at this time is that, according to the winit(v0.29) document, a rendering surface must be created after the Resumed event occurs, and it also states that when the Suspended event occurs, the existing rendering surface cannot be used.
Additionally, Android Studio's MainActivity had to inherit GameActivity, and full screen mode had to used to solve the swapchain size mismatch problem.